+++ module_adc/include/adc_service.h
@@ -71,19 +71,27 @@ typedef enum
     AD_7949_EXT_A0_P_EXT_A1_P=3
 } Ad7949ChannelInputs;
 
+
 /**
- * @brief Possible selection of analogue input pairs (to be sampled) in case of using AD7265 chip
- * used internally in AD7265 service.
+ * Structure type for channel mapping
  */
-typedef enum
+typedef struct
 {
-    AD_7265_CURRENT_B_C     =0,                     //corresponding to VA1 & VB1 in AD7265 (dc1k board)
-    AD_7265_VDC_IDC         =1,                     //corresponding to VA2 & VB2 in AD7265 (dc1k board)
-    AD_7265_AI_SIGNAL_1_3   =2,                     //corresponding to VA3 & VB3 in AD7265 (dc1k board)
-    AD_7265_AI_SIGNAL_2_4   =3,                     //corresponding to VA4 & VB4 in AD7265 (dc1k board)
-    AD_7265_BOARD_TEMP_PHASE_VOLTAGE_B  =4,         //corresponding to VA5 & VB5 in AD7265 (dc1k board)
-    AD_7265_PHASE_VOLTAGE_C_PHASE_VOLTAGE_B  =5     //corresponding to VA6 & VB6 in AD7265 (dc1k board)
-} Ad7265ChannelInputs;
+    int current_a;
+    int current_b;
+    int voltage_dc;
+    int current_dc;
+    int temperature;
+    int voltage_a;
+    int voltage_b;
+    int voltage_c;
+    int analogue_input_1;
+    int analogue_input_2;
+    int analogue_input_3;
+    int analogue_input_4;
+    int analogue_input_differential_mode_1;
+    int analogue_input_differential_mode_2;
+} Ad7265ChannelIndex;
 
 /**
  * @brief Fault Codes for sending to watchdog service
@@ -118,6 +126,7 @@ typedef struct
     out port ?p1_serial_clk;            /**< [[Nullable]] Port connecting to external ADC serial clock. */
     port ?p1_ready;                     /**< [[Nullable]] Port used to as ready signal for p32_adc_data ports and ADC chip. */
     out port ?p4_mux;                   /**< [[Nullable]] 4-bit Port used to control multiplexor on ADC chip. */
+    Ad7265ChannelIndex ad7265_channel_index;
 } AD7265Ports;
 
 /**
@@ -125,10 +134,12 @@ typedef struct
  */
 typedef struct
 {
+    int sign_phase_a;                   /**< Direction in which current on A Phase is measured [-1,1]. If 0, then phase A is not measured. */
     int sign_phase_b;                   /**< Direction in which current on B Phase is measured [-1,1]. */
     int sign_phase_c;                   /**< Direction in which current on C Phase is measured [-1,1]. */
     unsigned current_sensor_amplitude;  /**< Max amplitude of current the sensors that your DC board can handle. */
 }CurrentSensorsConfig;
+
 
 /**
  * Structure type for ports and configuration used by the ADC Service .