+++ module_watchdog/include/watchdog_service.h
@@ -14,14 +14,16 @@
  */
 typedef struct {
     out port ?p_shared_enable_tick_led; /**< 4-bit Port for Watchdog basic management. */
+    port ?p_diag_enable; /**< Nullable] Bidirectional port for Diagnostic and enable */
     out port ?p_tick; /**< [Nullable] Port for the periodic tick signal (if applicable in your SOMANET device). */
     port ?p_cpld_shared;
+    port ?p_cpld_fault_monitor;
 } WatchdogPorts;
 
 /**
  * @brief Enumeration of existing IFM modules.
  */
-enum {DC100_DC300, DC500, DC1K_DC5K};
+enum {DC100_DC300, DC500, DC1K_DC5K, DC1KD1, DC30};
 
 /**
  * @brief Service to manage the watchdog chip within your IFM SOMANET device.
@@ -40,7 +42,8 @@ void watchdog_service( WatchdogPorts &watchdog_ports, interface WatchdogInterfac
  * @param watchdog_ports Ports structure defining watchdog chip HW access.
  * @param IFM_module_type Type of the IFM board used to properly handle the ports.
  * @param output Value to be written to the shared LED port
+ * @param output_cpld Value to be written to the shared LED port for cpld boards
  * @param times Internal blinking counter
  * @param delay_counter External counter incrementing every WD clock cycle
  */
-void blink_red(int fault, int period, WatchdogPorts &watchdog_ports, int IFM_module_type, unsigned char &output, unsigned int &times, unsigned int &delay_counter);
+void blink_red(int fault, int period, WatchdogPorts &watchdog_ports, int IFM_module_type, unsigned char &output, unsigned &output_cpld, unsigned int &times, unsigned int &delay_counter);